+++ backend/server/src/models/app.rs
@@ -34,6 +34,7 @@ use sqlx::{Pool, Postgres};
 use std::env;
 use tonic::transport::Channel;
 use tower_http::cors::CorsLayer;
+use crate::spicedb::{check_permission, SpiceDbAuth};
 
 #[derive(Serialize)]
 pub enum AppMessage<T: Serialize> {
@@ -109,11 +110,50 @@ pub struct AppState {
     pub storage_bucket: Bucket,
     pub is_dev_env: bool,
     pub email_credentials: EmailCredentials,
-    /// Shared generated gRPC client for SpiceDB permission operations.
     pub spicedb: PermissionsServiceClient<Channel>,
-    /// Bearer key attached to authenticated SpiceDB requests.
     pub spicedb_key: String,
 }
+
+impl AppState {
+    /// Checks whether a user holds a permission on a SpiceDB resource, using
+    /// the application's shared SpiceDB client and credentials.
+    ///
+    /// Call this directly in handlers whose resource ID does not come from a
+    /// path parameter, for example when the ID is taken from the request body,
+    /// derived from a slug, or only known after a database lookup. When the
+    /// resource ID is a path parameter, prefer the [`SpiceDbAuth`] extractor.
+    ///
+    /// # Arguments
+    ///
+    /// * `user_id` - Chaos user to authorize
+    /// * `resource_type` - SpiceDB object type, such as `chaos/organisation`
+    /// * `resource_id` - Chaos ID of the resource, sent as the SpiceDB object ID
+    /// * `permission` - SpiceDB permission to check, such as `manage`
+    ///
+    /// # Returns
+    ///
+    /// * `Ok(())` if the user holds the permission
+    /// * `Err(ChaosError::ForbiddenOperation)` if the user does not
+    /// * `Err(ChaosError::InternalServerError)` if the SpiceDB call fails
+    pub async fn check_permission(
+        &self,
+        user_id: i64,
+        resource_type: &str,
+        resource_id: i64,
+        permission: &str,
+    ) -> Result<(), ChaosError> {
+        check_permission(
+            &self.spicedb,
+            &self.spicedb_key,
+            user_id,
+            resource_type,
+            resource_id,
+            permission,
+        )
+            .await
+    }
+}
+
 
 pub async fn init_app_state() -> AppState {
     // Initialise DB connection
@@ -169,8 +209,7 @@ pub async fn init_app_state() -> AppState {
     // Initialise email credentials
     let email_credentials = ChaosEmail::setup_credentials();
 
-    // Initialise the generated SpiceDB gRPC client. The lazy channel is shared
-    // by cheap client clones and connects when the first RPC is made.
+    // Initialise the generated SpiceDB gRPC client
     let spicedb_endpoint =
         env::var("SPICEDB_GRPC_ENDPOINT").expect("SPICEDB_GRPC_ENDPOINT must be set");
     let spicedb_key = env::var("SPICEDB_KEY").expect("SPICEDB_KEY must be set");
